+++ rollup.config.js
@@ -4,7 +4,7 @@ import shebang from 'rollup-plugin-shebang';
 import commonjs from 'rollup-plugin-commonjs';
 import resolve from 'rollup-plugin-node-resolve';
 
-export default [{
+const indexEntryPoints = [{
   input: 'index.js',
   output: {
     file: 'dist/index.js',
@@ -30,3 +30,44 @@ export default [{
     shebang(),
   ],
 }];
+
+const cleanupEntryPoints = [{
+  input: 'cleanup.js',
+  output: {
+    file: 'dist/cleanup.js',
+    format: 'cjs',
+  },
+  external: ['os', 'fs', 'path'],
+  plugins: [
+    resolve(),
+    commonjs(),
+    uglify({}, minify),
+  ],
+}, {
+  input: 'cleanup.js',
+  output: {
+    file: 'dist/cleanup-cli.js',
+    format: 'cjs',
+  },
+  external: ['os', 'fs', 'path'],
+  plugins: [
+    resolve(),
+    commonjs(),
+    uglify({}, minify),
+    shebang(),
+  ],
+}, {
+  input: 'cleanup-test.js',
+  output: {
+    file: 'dist/cleanup-test.js',
+    format: 'cjs',
+  },
+  external: ['os', 'fs', 'path', 'child_process'],
+  plugins: [
+    resolve(),
+    commonjs(),
+    uglify({}, minify),
+  ],
+}];
+
+export default [...indexEntryPoints, ...cleanupEntryPoints];