+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-_Nothing yet..._
+### Fixed
+- Correctly check global npm with `view version` check
 
 ## 1.0.4 - 2018-01-04
 
@@ -21,7 +22,7 @@ _Nothing yet..._
 ## 1.0.2 - 2017-11-14
 
 ### Fixed
-- Remove shebang from curl + eval use case 
+- Remove shebang from curl + eval use case
 - Add shebang to cli specific build
 
 ## 1.0.1 - 2017-11-14